requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
googletrans>=4.0.0
selenium>=4.15.0
aiohttp>=3.8.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0

# Search engines
googlesearch-python>=1.2.0
duckduckgo-search>=3.9.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
asyncio>=3.4.3
aiofiles>=23.0.0

# Additional utilities
urllib3>=2.0.0
lxml>=4.9.0

# AI and Machine Learning
google-generativeai>=0.3.2
